==> code/01_kaggle.ipynb <==
# Cleaning & Transforming 2022 Nurse Attrition Data
# Kim Bui & Camille Eastvold

## Source 2: CSV Dataset from Kaggle
https://www.kaggle.com/datasets/jpmiller/employee-attrition-for-healthcare?resource=download

import matplotlib.pyplot as plt
import seaborn as sns
import pandas as pd

### Load the dataset
1. find alike columns and drop others
2. total up percentages on each column based on if yes or no in attrition column 


# Load the dataset
watson_df=pd.read_csv('watson_healthcare_modified.csv')

display(watson_df)

#Drop unneeded columns
keep = ['Attrition','Age','Gender','JobSatisfaction','MaritalStatus','MonthlyIncome','OverTime']
watson_df=watson_df[keep]
display(watson_df)

# Change JobSatisfaction name and convert to 'Yes/No' category
watson_df['Job_Satisfaction'] = watson_df['JobSatisfaction'].apply(lambda x: 'Yes' if x > 2 else 'No')

# Optionally drop the original JobSatisfaction column if no longer needed
watson_df = watson_df.drop(columns=['JobSatisfaction'])

# Display the updated DataFrame
print("Updated DataFrame with Job_Satisfaction column modified:")
display(watson_df)


#Create New Income Column 
watson_df['Income'] = watson_df['MonthlyIncome'] * 12

# Drop the old 'MonthlyIncome' column
watson_df = watson_df.drop(columns=['MonthlyIncome'])

# Display the updated DataFrame
print("Updated DataFrame with Income column:")
display(watson_df)


attrition_counts = watson_df['Attrition'].value_counts()

# Display the counts
print("Total number of Yes and No in Attrition column:")
print(attrition_counts)


#Rename df
kaggle_df = watson_df

kaggle_df.to_csv("clean_kaggle.csv", index=False, encoding="utf-8")
